requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-groq
langchain-huggingface
faiss-cpu
pypdf
python-dotenv
streamlit
requests
